Fix build issue for non-newlib build introduced with a34a8f3e19

+++ src/syscalls/interfaces/mod.rs
@@ -393,7 +393,7 @@ pub trait SyscallInterface: Send + Sync {
 	}
 
 	#[cfg(not(feature = "newlib"))]
-	fn stat(&self, _file: *const u8, _st: usize) -> isize {
+	fn stat(&self, _file: *const u8, _st: usize) -> i32 {
 		info!("stat is unimplemented");
 		-ENOSYS
 	}
